Add validation to attribute approver to only allow configuring plugins that have been registered
