+++ pkg/approver/fake/webhook.go
@@ -25,8 +25,8 @@ import (
 
 var _ approver.Webhook = &FakeWebhook{}
 
-// FakeWebhook is a testing webook designed to mock webhooks with a determined
-// response.
+// FakeWebhook is a testing webook designed to mock webhooks with a
+// pre-determined response.
 type FakeWebhook struct {
 	validateFunc func(context.Context, *cmpapi.CertificateRequestPolicy) (approver.WebhookValidationResponse, error)
 }
